+++ App.cpp
@@ -1,8 +1,9 @@
 # include <iostream>
 # include <string>
 # include <fstream>
-# include "Recipes.h" // Include the header file that contains the definition of the 'Recipes' class
-# include "Ingredients.h" // Include the header file that contains the definition of the 'Ingredients' class
+# include "./src/Recipes.cpp" // Include the header file that contains the definition of the 'Recipes' class
+# include "./src/Ingredients.cpp" // Include the header file that contains the definition of the 'Ingredients' class
+# include "./src/Support.cpp" // Include the header file that contains the definition of the 'Ingredients' class
 # include <array>
 # include <sstream>
 # include <iterator>
@@ -19,253 +20,9 @@ using namespace std;
    @author Cooks 
 */
 
-//Enum of values of the difficulty levels
-enum Level {
-    None = 0,
-    Easy = 1,
-    Medium = 2,
-    Hard = 3
-};
-
-// Function to split a string into a vector of strings
-vector<string> split_string(string str, string delimiter, bool lowercase = true){
-    // lowercase the string
-    if (lowercase){
-        transform(str.begin(), str.end(), str.begin(), ::tolower);
-    }
-    vector<string> result;
-    size_t pos = 0;
-    string token;
-    while ((pos = str.find(delimiter)) != string::npos){
-        token = str.substr(0, pos);
-        // remove leading and trailing spaces
-        token.erase(0, token.find_first_not_of(" "));
-        token.erase(token.find_last_not_of(" ") + 1);
-
-        result.push_back(token);
-        str.erase(0, pos + delimiter.length());
-    }
-    str.erase(0, str.find_first_not_of(" "));
-    str.erase(str.find_last_not_of(" ") + 1);
-    result.push_back(str);
-    return result;
-}
-
-    /***************************************************************************************************************************************
-     *          Ingredients Class                                                                                                          *
-    ****************************************************************************************************************************************/
-
-    /**Implementation of Ingredients functions */
-    /*Default Constructor */
-    Ingredients::Ingredients(){
-        quantity = 0;
-    };
-
-    /*Defined Constructor*/
-    Ingredients::Ingredients(double q, string un, string nm) {
-        name = nm;
-        quantity = q;
-        unit = un;
-    }
-
-    /*Destructor code (idk what this is for) */
-    Ingredients::~Ingredients() {
-        // Destructor code here
-    }
-
-    /* Accessor function to get the name of the ingredient
-        @return name of the ingredient
-    */
-    string Ingredients::getName() {
-        return name;
-    }
-
-    /* Accessor function to get the unit of the ingredient
-        @return unit of the ingredient
-    */
-    string Ingredients::getUnit() {
-        return unit;
-    }
-
-    /* Accessor function to get the quantity of the ingredient
-        @return quantity of the ingredient
-    */
-    double Ingredients::getQuantity() {
-        return quantity;
-    }
-
-    /* Function to return a string representation of the ingredient
-        @return string representation of the ingredient
-    */
-    string Ingredients::toString() {
-        string result = "";
-        if (quantity != 0){
-            result += to_string(quantity);
-            if (result.find('.') != string::npos){
-                result.erase(result.find_last_not_of('0') + 1, string::npos);
-                if (result.back() == '.'){
-                    result.pop_back();
-                }
-            }
-            result += " ";
-        }
-        if (unit != "null") {
-            result += unit + " ";
-        }
-        result += name;
-        return result;
-    }
-
-
-    /***************************************************************************************************************************************
-     *          Recipes Class                                                                                                              *
-    ****************************************************************************************************************************************/
-
-    /**Implementation of Recipes functions */
-    /*Default Constructor */
-    Recipes::Recipes(){
-        time = 0;
-        difficulty = None;
-    };
-
-    /*Defined Constructor*/
-    Recipes::Recipes(string nm, string ser, string desc, vector<Ingredients> ing, string ins, int t, enum Level lv) {
-        name = nm;
-        series = ser;
-        description = desc;
-        ingredients = ing;
-
-        ingredients_str = "";
-        for (int i = 0; i < ingredients.size(); i++){
-            ingredients_str += "- " +ingredients[i].toString();
-            if (i != ingredients.size() - 1){
-                ingredients_str += "\n";
-            }
-        }
-        ingredient_names = {};
-        for (int i = 0; i < ingredients.size(); i++){
-            string ingredient_name = ingredients[i].getName();
-            // make lowercase and remove leading and trailing spaces
-            transform(ingredient_name.begin(), ingredient_name.end(), ingredient_name.begin(), ::tolower);
-            ingredient_name.erase(0, ingredient_name.find_first_not_of(" "));
-            ingredient_name.erase(ingredient_name.find_last_not_of(" ") + 1);
-
-            ingredient_names.push_back( ingredient_name );
-        }
-        sort(ingredient_names.begin(), ingredient_names.end());
-
-        instructions = ins;
-        time = t;
-        difficulty = lv;
-
-        
-    }
-
-    /*Destructor code (idk what this is for)*/
-    Recipes::~Recipes() {
-        // Destructor code here
-    }
-
-    /* Accessor function to get the name of the recipe
-        @return name of the recipe
-    */
-    string Recipes::getName() {
-        return name;
-    }
-
-    /* Accessor function to get the series of the recipe
-        @return series of the recipe
-    */
-    string Recipes::getSeries() {
-        return series;
-    }
-
-    /* Accessor function to get the description of the recipe
-        @return short description of the recipe
-    */
-    string Recipes::getDescription() {
-        return description;
-    }
-
-    /* Accessor function to get the ingredients of the recipe
-        @return ingredients of the recipe (list)
-    */
-    vector<Ingredients> Recipes::getIngredients() {
-        return ingredients;
-    }
-
-    /* Accessor function to get the ingredients of the recipe
-        @return ingredients of the recipe (list)
-    */
-    vector<string> Recipes::getIngredientsNames() {
-        return ingredient_names;
-    }
-
-    /* Accessor function to get the instructions of the recipe
-        @return instructions of the recipe
-    */
-    string Recipes::getInstructions() {
-        return instructions;
-    }
-
-    /* Accessor function to get the time of the recipe
-        @return time of the recipe in minutes
-    */
-    int Recipes::getTime() {
-        return time;
-    }
-
-    /* Accessor function to get the difficulty of the recipe
-        @return difficulty of the recipe as an int (1, 2, 3)
-    */
-    enum Level Recipes::getDifficulty() {
-        return difficulty;
-    }
-
-    /* Function to return a string representation of the recipe
-        @return string representation of the recipe
-    */
-    string Recipes::toString() {
-        string result = "Name: " + name + "\n";
-        result += "Series: " + series + "\n";
-        result += "Description: " + description + "\n";
-        result += "Ingredients: \n" + ingredients_str + "\n";
-        result += "Instructions: " + instructions + "\n";
-        result += "Time: " + to_string(time) + " minutes\n";
-        result += "Difficulty: ";
-
-        if (difficulty == 1){
-            result += "Easy\n";
-        }
-        else if (difficulty == 2){
-            result += "Medium\n";
-        }
-        else if (difficulty == 3){
-            result += "Hard\n";
-        }
-        return result;
-    }    
-
-    /*Function to return a string of the recipe name, series, time, and difficulty
-        @return string of the recipe name, series, time, and difficulty
-    */
-    string Recipes::toStringShort() {
-        string result = name + " (" + series + ") - " + to_string(time) + " minutes - ";
-        if (difficulty == 1){
-            result += "Easy";
-        }
-        else if (difficulty == 2){
-            result += "Medium";
-        }
-        else if (difficulty == 3){
-            result += "Hard";
-        }
-        return result;
-    }
-
-    /***************************************************************************************************************************************
-     *          Setup functions                                                                                                            *
-    ****************************************************************************************************************************************/
+/***************************************************************************************************************************************
+ *          Setup functions                                                                                                            *
+****************************************************************************************************************************************/
 
 
 /* Vector of the recipes, declared based on number of recipes in the database
@@ -309,7 +66,7 @@ vector<Ingredients> process_ingred(string ingred){
     return ingredients_list;
 }
 
-vector<string> process_instructions(string instruct){
+string process_instructions(string instruct){
     string instructions;
     // if the start and end of the string are both " then remove them
     if (instruct.front() == '"' && instruct.back() == '"'){
@@ -322,14 +79,9 @@ vector<string> process_instructions(string instruct){
         instructions = instruct;
     }
 
-    vector<string> instructions_parts = split_string(instructions, "/", false);
-    for (string instruction : instructions_parts){
-        if (instruction.empty()){
-            // remove empty instructions
-            continue;
-        }
-    }
-    return instructions_parts;
+    instructions = replaceAll(instructions, "/", "\n");
+
+    return instructions;
 
 }
 
@@ -384,7 +136,7 @@ void setup(string fileName){
             continue;
         }
 
-        string name, series, description;
+        string name, series, description, instructions;
         int time;
         enum Level difficulty;
 
@@ -395,11 +147,11 @@ void setup(string fileName){
         series = parts[1];
         description = parts[2];
         vector<Ingredients> ingredients_parts = process_ingred(parts[3]);
-        vector<string> instructions_parts = process_instructions(parts[4]);
+        instructions = process_instructions(parts[4]);
         time = process_time(parts[5]);
         difficulty = process_difficulty(parts[6]);
 
-        Recipes newRecipe = Recipes(name, series, description, ingredients_parts, parts[4], time, difficulty);
+        Recipes newRecipe = Recipes(name, series, description, ingredients_parts, instructions, time, difficulty);
         allRecipes.push_back(newRecipe);
         if (find(allSeries.begin(), allSeries.end(), series) == allSeries.end()){
             allSeries.push_back(series);
@@ -408,26 +160,9 @@ void setup(string fileName){
     }
 }
 
-    /***************************************************************************************************************************************
-     *          Supporting functions                                                                                                       *
-    ****************************************************************************************************************************************/
-
-
-/* Function to replace comma space to comma
-   Remove empty spaces
-*/
-string replaceAll(string str, string from, string to){
-    // check if from and to empty
-    if (from.empty() || to.empty()){
-        return str;
-    }
-    size_t start_pos = 0;
-    while ((start_pos = str.find(from, start_pos)) != string::npos){
-        str.replace(start_pos, from.length(), to);
-        start_pos += to.length();
-    }
-    return str;
-}
+/***************************************************************************************************************************************
+ *          Options functions                                                                                                          *
+****************************************************************************************************************************************/
 
 /* Function to display recipes specified in the param
     @param array of recipes
@@ -436,13 +171,6 @@ void display(vector<Recipes> recipes){
     for (int i = 0; i < recipes.size(); i++){
         cout << recipes[i].toStringShort() << endl;
     }
-}
-
-/* Function to quit the program
-*/
-void quit(){
-    cout << "Goodbye!" << endl;
-    exit(0);
 }
 
 /* Function to sort the recipes by difficulty
@@ -463,20 +191,10 @@ bool sortByTime(Recipes a, Recipes b){
     return a.getTime() < b.getTime();
 }
 
-/* Function to read the whole line of user input instead of just one word
- */
-string readLine(bool ignore = true){
-    string s;
-    if (ignore){
-        cin.ignore();
-    }
-    getline(cin, s);
-    return s;
-}
 
-    /***************************************************************************************************************************************
-     *          Search functions                                                                                                           *
-    ****************************************************************************************************************************************/
+/***************************************************************************************************************************************
+ *          Search functions                                                                                                           *
+****************************************************************************************************************************************/
 
 /* Function to search the recipes array by specific name
     Prints out recipes with the searched string in its name
@@ -617,9 +335,9 @@ vector<Recipes> searchBySeries(){
 }
 
 
-    /***************************************************************************************************************************************
-     *          Set filters and random                                                                                                     *
-    ****************************************************************************************************************************************/
+/***************************************************************************************************************************************
+ *          Set filters and random                                                                                                     *
+****************************************************************************************************************************************/
 
 /* Function to set filters for the recipes
     Filters the recipes based on the user input
@@ -771,9 +489,9 @@ void random(){
     cout << allRecipes[randomIndex].toString() << endl;
 }
 
-    /***************************************************************************************************************************************
-     *          Options and main                                                                                                           *
-    ****************************************************************************************************************************************/
+/***************************************************************************************************************************************
+ *          Options and main                                                                                                           *
+****************************************************************************************************************************************/
 
 /* Function to display the options for the user
     Takes the user input and calls the appropriate function
